add test_network_utils.py
